+++ CHANGELOG.md
@@ -50,6 +50,7 @@ _This changelog only relates to the `raffia` crate._
 - Added support of allowing keyframe block in Sass `@mixin` and `@include` at-rule.
 - Added support of allowing Sass at-rule in `@keyframes` at-rule.
 - Fixed delimiter being calculated in Sass binary expression.
+- Added support of deprecated Sass `@import` syntax.
 
 ## v0.2.0